hope to be in right direction
