+++ app/src/main/res/layout/fragment_home.xml
@@ -7,8 +7,12 @@
 
     <indoormaps.ninepatch.com.library.IndoorMapsView
             android:id="@+id/map_view"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent" />
+            android:layout_width="0dp"
+            android:layout_height="match_parent"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent" />
 
     <!--    <FrameLayout-->
     <!--            android:layout_width="wrap_content"-->
