+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -6,10 +6,10 @@
     app:startDestination="@+id/navigation_home">
 
     <fragment
-        android:id="@+id/navigation_home"
-        android:name="com.example.indoorapplication.ui.home.HomeFragment"
-        android:label="@string/title_home"
-        tools:layout="@layout/fragment_home" >
+            android:id="@+id/navigation_home"
+            android:name="com.example.indoorapplication.ui.settings.SettingsFragment"
+            android:label="@string/title_home"
+            tools:layout="@layout/fragment_settings">
         <action
                 android:id="@+id/action_navigation_home_to_navigation_dashboard"
                 app:destination="@id/navigation_dashboard" />
